Moved channel playlist tests
