+++ Tribler/Test/Core/Modules/RestApi/test_my_channel_endpoints.py
@@ -228,44 +228,3 @@ class TestMyChannelEndpoints(AbstractTestMyChannelEndpoints):
         self.insert_torrents_into_my_channel(torrent_list)
 
         return self.do_request('mychannel/torrents', expected_code=200).addCallback(verify_torrents_json)
-
-
-class TestMyChannelPlaylistEndpoints(AbstractTestMyChannelEndpoints):
-
-    def create_playlist(self, channel_id, dispersy_id, peer_id, name, description):
-        self.channel_db_handler.on_playlist_from_dispersy(channel_id, dispersy_id, peer_id, name, description)
-
-    def insert_torrent_into_playlist(self, playlist_disp_id, infohash):
-        self.channel_db_handler.on_playlist_torrent(42, playlist_disp_id, 42, infohash)
-
-    @deferred(timeout=10)
-    def test_get_playlists_endpoint_without_channel(self):
-        """
-        Testing whether the API returns error 404 if no channel has been created when fetching playlists
-        """
-        self.should_check_equality = False
-        return self.do_request('mychannel/playlists', expected_code=404)
-
-    @deferred(timeout=10)
-    def test_playlists_endpoint_no_playlists(self):
-        """
-        Testing whether the API returns the right JSON data if no playlists have been added to your channel
-        """
-        self.create_my_channel("my channel", "this is a short description")
-        return self.do_request('mychannel/playlists', expected_code=200, expected_json={"playlists": []})
-
-    @deferred(timeout=10)
-    def test_playlists_endpoint(self):
-        """
-        Testing whether the API returns the right JSON data if playlists are fetched
-        """
-        my_channel_id = self.create_my_channel("my channel", "this is a short description")
-        self.create_playlist(my_channel_id, 1234, 42, "test playlist", "test description")
-        torrent_list = [[my_channel_id, 1, 1, 'a' * 20, 1460000000, "ubuntu-torrent.iso", [], []]]
-        self.insert_torrents_into_my_channel(torrent_list)
-        self.insert_torrent_into_playlist(1234, 'a' * 20)
-
-        expected_json = {u"playlists": [{u"id": 1, u"name": u"test playlist", u"description": u"test description",
-                                         u"torrents": [{u"infohash": bytes(('a' * 20).encode('hex')),
-                                                        u"name": u"ubuntu-torrent.iso"}]}]}
-        return self.do_request('mychannel/playlists', expected_code=200, expected_json=expected_json)
